+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+- **CSD-103**
+  - Fixed the getExecuteOptions select handler in the granule detail view modal to correctly set the workflow state data.
+
 ### Breaking Changes
 
 ### Added